New pairing pipeline work-in-progress

+++ colabfold/mmseqs/search.py
@@ -1,5 +1,6 @@
 """
 Functionality for running mmseqs locally. Takes in a fasta file, outputs final.a3m
+Pairing for homodimer prediction, to get the best hit per species
 """
 
 import logging
@@ -215,6 +216,7 @@ def mmseqs_search_pair(
     spire_db: Path = Path("spire_ctg10_2401_db"),
     mmseqs: Path = Path("mmseqs"),
     pair_env: bool = True,
+    filter: bool = False,
     prefilter_mode: int = 0,
     s: float = 8,
     threads: int = 64,
@@ -262,15 +264,16 @@ def mmseqs_search_pair(
     if gpu_server:
         search_param += ["--gpu-server", str(gpu_server)]
     expand_param = ["--expansion-mode", "0", "-e", "inf", "--expand-filter-clusters", "0", "--max-seq-id", "0.95",]
+    filter_param = ["--filter-msa", str(1 if filter else 0), "--filter-min-enable", "1000", "--diff", "3000", "--qid", "0.2,0.4,0.6,0.8,1.0", "--qsc", "0", "--max-seq-id", "0.95",]
     run_mmseqs(mmseqs, ["search", base.joinpath("qdb"), dbbase.joinpath(db), base.joinpath("res"), base.joinpath("tmp"), "--threads", str(threads),] + search_param,)
     run_mmseqs(mmseqs, ["mvdb", base.joinpath("tmp/latest/profile_1"), base.joinpath("prof_res")])
     run_mmseqs(mmseqs, ["lndb", base.joinpath("qdb_h"), base.joinpath("prof_res_h")])    
     run_mmseqs(mmseqs, ["expandaln", base.joinpath("qdb"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res"), dbbase.joinpath(f"{db}{dbSuffix2}"), base.joinpath("res_exp"), "--db-load-mode", str(db_load_mode), "--threads", str(threads),] + expand_param,)
-    run_mmseqs(mmseqs, ["align", base.joinpath("prof_res"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res_exp"), base.joinpath("res_exp_realign"), "--db-load-mode", str(db_load_mode), "-e", "0.001", "--max-accept", "1000000", "--threads", str(threads),],)
+    run_mmseqs(mmseqs, ["align", base.joinpath("prof_res"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res_exp"), base.joinpath("res_exp_realign"), "--db-load-mode", str(db_load_mode), "--alignment-mode", "1", "-e", "0.001", "--max-accept", "1000000", "--threads", str(threads),],)  # hack
     run_mmseqs(mmseqs, ["pairaln", base.joinpath("qdb"), dbbase.joinpath(f"{db}"), base.joinpath("res_exp_realign"), base.joinpath("res_exp_realign_pair"), "--db-load-mode", str(db_load_mode), "--pairing-mode", str(pairing_strategy), "--pairing-dummy-mode", "0", "--threads", str(threads), ],)
     run_mmseqs(mmseqs, ["align", base.joinpath("prof_res"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res_exp_realign_pair"), base.joinpath("res_exp_realign_pair_bt"), "--db-load-mode", str(db_load_mode), "-e", "inf", "-a", "--threads", str(threads), ],)
     run_mmseqs(mmseqs, ["pairaln", base.joinpath("qdb"), dbbase.joinpath(f"{db}"), base.joinpath("res_exp_realign_pair_bt"), base.joinpath("res_final"), "--db-load-mode", str(db_load_mode), "--pairing-mode", str(pairing_strategy), "--pairing-dummy-mode", "1", "--threads", str(threads),],)
-    run_mmseqs(mmseqs, ["result2msa", base.joinpath("qdb"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res_final"), base.joinpath("pair.a3m"), "--db-load-mode", str(db_load_mode), "--msa-format-mode", "5", "--threads", str(threads),],)
+    run_mmseqs(mmseqs, ["result2msa", base.joinpath("qdb"), dbbase.joinpath(f"{db}{dbSuffix1}"), base.joinpath("res_final"), base.joinpath("pair.a3m"), "--db-load-mode", str(db_load_mode),  "--msa-format-mode", "5", "--threads", str(threads),] + filter_param,)
     if unpack:
         run_mmseqs(mmseqs, ["unpackdb", base.joinpath("pair.a3m"), base.joinpath("."), "--unpack-name-mode", "0", "--unpack-suffix", output,],)
         run_mmseqs(mmseqs, ["rmdb", base.joinpath("pair.a3m")])
@@ -342,8 +345,8 @@ def main():
         "--filter",
         type=int,
         default=1,
-        choices=[0, 1],
-        help="Filter the MSA by pre-defined align_eval, qsc, max_accept",
+        choices=[0, 1, 2],
+        help="Filter the MSA by pre-defined align_eval, qsc, max_accept; 2: also filter paired MSA",
     )
 
     # mmseqs params
@@ -381,6 +384,13 @@ def main():
         help="align - Maximum accepted alignments before alignment calculation for a query is stopped.",
     )
     parser.add_argument(
+        "--pair-mode",
+        help="Multimer MSA pairing mode for complex prediction: unpaired MSA only, paired MSA only, both (default).",
+        type=str,
+        default="unpaired_paired",
+        choices=["unpaired", "paired", "unpaired_paired"],
+    )
+    parser.add_argument(
         "--pairing_strategy", type=int, default=0, help="pairaln - Pairing strategy."
     )
     parser.add_argument(
@@ -414,9 +424,9 @@ def main():
     logging.basicConfig(level = logging.INFO)
 
     queries, is_complex = get_queries(args.query, None)
-
+    is_complex = True # hack
     queries_unique = []
-    for job_number, (raw_jobname, query_sequences, a3m_lines, other_molecules) in enumerate(queries):
+    for job_number, (raw_jobname, query_sequences, _, other_molecules) in enumerate(queries):
         # remove duplicates before searching
         query_sequences = (
             [query_sequences] if isinstance(query_sequences, str) else query_sequences
@@ -465,35 +475,47 @@ def main():
                 id += 1
             file_number += 1
 
-    mmseqs_search_monomer(
-        mmseqs=args.mmseqs,
-        dbbase=args.dbbase,
-        base=args.base,
-        uniref_db=args.db1,
-        template_db=args.db2,
-        metagenomic_db=args.db3,
-        use_env=args.use_env,
-        use_templates=args.use_templates,
-        filter=args.filter,
-        expand_eval=args.expand_eval,
-        align_eval=args.align_eval,
-        diff=args.diff,
-        qsc=args.qsc,
-        max_accept=args.max_accept,
-        prefilter_mode=args.prefilter_mode,
-        s=args.s,
-        db_load_mode=args.db_load_mode,
-        threads=args.threads,
-        gpu=args.gpu,
-        gpu_server=args.gpu_server,
-        unpack=args.unpack,
-    )
-    if is_complex is True:
+    keep_paired = args.pair_mode == "paired" or args.pair_mode == "unpaired_paired"
+    keep_unpaired = args.pair_mode == "unpaired" or args.pair_mode == "unpaired_paired"
+
+    if keep_unpaired:
+        mmseqs_search_monomer(
+            mmseqs=args.mmseqs,
+            dbbase=args.dbbase,
+            base=args.base,
+            uniref_db=args.db1,
+            template_db=args.db2,
+            metagenomic_db=args.db3,
+            use_env=args.use_env,
+            use_templates=args.use_templates,
+            filter=args.filter > 0,
+            expand_eval=args.expand_eval,
+            align_eval=args.align_eval,
+            diff=args.diff,
+            qsc=args.qsc,
+            max_accept=args.max_accept,
+            prefilter_mode=args.prefilter_mode,
+            s=args.s,
+            db_load_mode=args.db_load_mode,
+            threads=args.threads,
+            gpu=args.gpu,
+            gpu_server=args.gpu_server,
+            unpack=args.unpack,
+        )
+    else:
+        id = 0
+        for job_number, (raw_jobname, query_sequences, query_seqs_cardinality, other_molecules) in enumerate(queries_unique):
+            for seq in query_sequences:
+                args.base.joinpath(f"{id}.a3m").touch()
+                id += 1
+
+    if is_complex is True and keep_paired:
         mmseqs_search_pair(
             mmseqs=args.mmseqs,
             dbbase=args.dbbase,
             base=args.base,
             uniref_db=args.db1,
+            filter=args.filter == 2,
             prefilter_mode=args.prefilter_mode,
             s=args.s,
             db_load_mode=args.db_load_mode,
@@ -511,6 +533,7 @@ def main():
                 base=args.base,
                 uniref_db=args.db1,
                 spire_db=args.db4,
+                filter=args.filter == 2,
                 prefilter_mode=args.prefilter_mode,
                 s=args.s,
                 db_load_mode=args.db_load_mode,
@@ -522,6 +545,7 @@ def main():
                 unpack=args.unpack,
             )
 
+    if is_complex is True:
         if args.unpack or args.af3_json:
             id = 0
             for job_number, (
@@ -532,27 +556,28 @@ def main():
             ) in enumerate(queries_unique):
                 unpaired_msa = []
                 paired_msa = None
-                if len(query_seqs_cardinality) > 1:
+                if len(query_seqs_cardinality) >= 1: # hack
                     paired_msa = []
                 for seq in query_sequences:
-                    with args.base.joinpath(f"{id}.a3m").open("r") as f:
-                        unpaired_msa.append(f.read())
-                    if args.unpack:
-                        args.base.joinpath(f"{id}.a3m").unlink()
-
-                    if args.use_env_pairing:
-                        with open(args.base.joinpath(f"{id}.paired.a3m"), 'a') as file_pair:
-                            with open(args.base.joinpath(f"{id}.env.paired.a3m"), 'r') as file_pair_env:
-                                while chunk := file_pair_env.read(10 * 1024 * 1024):
-                                    file_pair.write(chunk)
+                    if keep_unpaired:
+                        with args.base.joinpath(f"{id}.a3m").open("r") as f:
+                            unpaired_msa.append(f.read())
                         if args.unpack:
-                            args.base.joinpath(f"{id}.env.paired.a3m").unlink()
+                            args.base.joinpath(f"{id}.a3m").unlink()
+                    if keep_paired:
+                        if args.use_env_pairing:
+                            with open(args.base.joinpath(f"{id}.paired.a3m"), 'a') as file_pair:
+                                with open(args.base.joinpath(f"{id}.env.paired.a3m"), 'r') as file_pair_env:
+                                    while chunk := file_pair_env.read(10 * 1024 * 1024):
+                                        file_pair.write(chunk)
+                            if args.unpack:
+                                args.base.joinpath(f"{id}.env.paired.a3m").unlink()
 
-                    if len(query_seqs_cardinality) > 1:
-                        with args.base.joinpath(f"{id}.paired.a3m").open("r") as f:
-                            paired_msa.append(f.read())
-                    if args.unpack:
-                        args.base.joinpath(f"{id}.paired.a3m").unlink()
+                        if paired_msa is not None:
+                            with args.base.joinpath(f"{id}.paired.a3m").open("r") as f:
+                                paired_msa.append(f.read())
+                        if args.unpack:
+                            args.base.joinpath(f"{id}.paired.a3m").unlink()
                     id += 1
 
                 if args.af3_json:
